Fix iconv breakage
Stops the following test failure:

    not ok 13 iconv works
    # (in test file common/php.bats, line 105)
    #   `[ "$output" = "foobar" ]' failed
    # status: 0
    # output: PHP Warning:  iconv(): Wrong encoding, conversion from "UTF-8" to "ASCII//TRANSLIT" is not allowed in Command line code on line 1

The edge version of gnu-libiconv is now at 1.16 and this version no longer
includes the preload library which is needed to make iconv work properly.

https://gitlab.alpinelinux.org/alpine/aports/-/issues/12328

Alpine 3.10 is the first version that contains gnu-libiconv with the preload library.
For images based on <3.10, they install it separately from 3.10.
For images based on >=3.10, they install it as part of the regular package installation.